Finish header content in html & update README
Finished up the header content in the html file. I'll probably have to
change some of it later on when I start styling it, but for now I'm okay
with how it is going.
Also, added text to give credit to the creators of the images I'm going
to be using for the page layout.