Style header text and image
I styled the header so that each section takes up 50% of the content area.
I had some problems resizing the image - finally decided I should set
max-height for the div container and then added object-fit property to the
image itself. I don't really like how the picture display area changes
with different screen sizes. Will have to find a solution to this problem.

I also changed the padding of the navbar and added <a> to the "button"
element in the header section.