Finish header section
I *kind of* simplified the code for the header image, for which I had to
edit the html file as well. I'm still not completely satisfied with the
outcome, as I don't really like the resizing of the image when it shrinks...

I styled the button - probably not the most graceful solution, I should
have just marked it as a <button> element in the html file instead of
<span>. Maybe I'll change it later.

I've added some hovering effect to the navbar links and the signup button
in the header section.